Move more Hive tests
